Reverted Aaron's erroneous changes in AnimationScrollDatagrid. Fixed small warnings.

+++ src/Orbiter/DesignObj_DataGrid.cpp
@@ -198,7 +198,7 @@ g_PlutoProfiler->Stop("send command");
 						{
 							size_t size;
 							pCell->m_pGraphicData = FileUtils::ReadFileIntoBuffer(pPath,size);
-							pCell->m_GraphicLength=size;
+							pCell->m_GraphicLength=(unsigned long)size;
 						}
 						else
 						{
